@@ -4,7 +4,7 @@ source 'https://rubygems.org'
 # This can be removed once all gems depend on > 3.0
 gem 'mime-types', '~> 2.99', require: 'mime/types/columnar'
 
-gem 'rails', '~> 4.2.7'
+gem 'rails', '~> 5.0.0'
 gem 'rails-i18n'
 
 gem 'autoprefixer-rails'
@@ -16,7 +16,7 @@ gem 'daemons'
 gem 'dalli'
 gem 'delayed_job'
 gem 'delayed_job_active_record'
-gem 'doorkeeper'
+gem 'doorkeeper', '~> 4.0.0.rc4'
 gem 'dynamic_form'
 gem 'gchartrb', require: 'google_chart'
 gem 'gravtastic'
@@ -28,7 +28,7 @@ gem 'jquery-rails'
 gem 'mail'
 gem 'multi_json'
 gem 'newrelic_rpm'
-gem 'paul_revere', '~> 2.0'
+gem 'paul_revere', '~> 2.1.0.rc1'
 gem 'pg'
 gem 'psych', '~> 2.0.12'
 gem 'rack'
