@@ -16,9 +16,9 @@ class Api::V1::OwnersController < Api::BaseController
     owner = User.find_by_name(params[:email])
     if owner
       @rubygem.ownerships.create(user: owner)
-      render text: 'Owner added successfully.'
+      render plain: 'Owner added successfully.'
     else
-      render text: 'Owner could not be found.', status: :not_found
+      render plain: 'Owner could not be found.', status: :not_found
     end
   end
 
@@ -26,12 +26,12 @@ class Api::V1::OwnersController < Api::BaseController
     owner = @rubygem.owners.find_by_name(params[:email])
     if owner
       if @rubygem.ownerships.find_by_user_id(owner.id).try(:destroy)
-        render text: "Owner removed successfully."
+        render plain: "Owner removed successfully."
       else
-        render text: 'Unable to remove owner.', status: :forbidden
+        render plain: 'Unable to remove owner.', status: :forbidden
       end
     else
-      render text: 'Owner could not be found.', status: :not_found
+      render plain: 'Owner could not be found.', status: :not_found
     end
   end
 
@@ -44,7 +44,7 @@ class Api::V1::OwnersController < Api::BaseController
         format.yaml { render yaml: rubygems }
       end
     else
-      render text: "Owner could not be found.", status: :not_found
+      render plain: "Owner could not be found.", status: :not_found
     end
   end
 
@@ -52,6 +52,6 @@ class Api::V1::OwnersController < Api::BaseController
 
   def verify_gem_ownership
     return if current_user.rubygems.find_by_name(params[:rubygem_id])
-    render text: 'You do not have permission to manage this gem.', status: :unauthorized
+    render plain: 'You do not have permission to manage this gem.', status: :unauthorized
   end
 end
