@@ -1,9 +1,9 @@
 class Api::V1::Versions::DownloadsController < Api::BaseController
   def index
-    render text: "This endpoint is not supported anymore", status: :gone
+    render plain: "This endpoint is not supported anymore", status: :gone
   end
 
   def search
-    render text: "This endpoint is not supported anymore", status: :gone
+    render plain: "This endpoint is not supported anymore", status: :gone
   end
 end
