@@ -1,11 +1,11 @@
 class Internal::PingController < ApplicationController
   def index
-    ActiveRecord::Base.connection.select_value('SELECT 1') == '1' or \
+    ActiveRecord::Base.connection.select_value('SELECT 1') == 1 or \
       raise StandardError, 'Failed to SELECT 1 from DB server'
-    render text: 'PONG'
+    render plain: 'PONG'
   end
 
   def revision
-    render text: AppRevision.version
+    render plain: AppRevision.version
   end
 end
