@@ -106,7 +106,7 @@ class Rubygem < ActiveRecord::Base
   end
 
   def public_versions_with_extra_version(extra_version)
-    versions = public_versions(5)
+    versions = public_versions(5).to_a
     versions << extra_version
     versions.uniq.sort_by(&:position)
   end
