@@ -17,7 +17,7 @@ class Api::V1::DeletionsControllerTest < ActionController::TestCase
 
       context "ON DELETE to create for existing gem version" do
         setup do
-          delete :create, gem_name: @rubygem.to_param, version: @v1.number
+          delete :create, params: { gem_name: @rubygem.to_param, version: @v1.number }
         end
         should respond_with :success
         should "keep the gem, deindex, keep owner" do
@@ -38,7 +38,7 @@ class Api::V1::DeletionsControllerTest < ActionController::TestCase
 
         context "ON DELETE to create for version 0.1.1" do
           setup do
-            delete :create, gem_name: @rubygem.to_param, version: @v2.number
+            delete :create, params: { gem_name: @rubygem.to_param, version: @v2.number }
           end
           should respond_with :success
           should "keep the gem, deindex it, and keep the owners" do
@@ -61,7 +61,7 @@ class Api::V1::DeletionsControllerTest < ActionController::TestCase
 
         context "ON DELETE to create for version 0.1.1 and x86-darwin-10" do
           setup do
-            delete :create, gem_name: @rubygem.to_param, version: @v2.number, platform: @v2.platform
+            delete :create, params: { gem_name: @rubygem.to_param, version: @v2.number, platform: @v2.platform }
           end
           should respond_with :success
           should "keep the gem, deindex it, and keep the owners" do
@@ -85,7 +85,7 @@ class Api::V1::DeletionsControllerTest < ActionController::TestCase
 
       context "ON DELETE to create for existing gem with invalid version" do
         setup do
-          delete :create, gem_name: @rubygem.to_param, version: "0.2.0"
+          delete :create, params: { gem_name: @rubygem.to_param, version: "0.2.0" }
         end
         should respond_with :not_found
         should "not modify any versions" do
@@ -101,7 +101,7 @@ class Api::V1::DeletionsControllerTest < ActionController::TestCase
         setup do
           @other_user = create(:user)
           @request.env["HTTP_AUTHORIZATION"] = @other_user.api_key
-          delete :create, gem_name: @rubygem.to_param, version: '0.1.0'
+          delete :create, params: { gem_name: @rubygem.to_param, version: '0.1.0' }
         end
         should respond_with :forbidden
         should "not record the deletion" do
@@ -112,7 +112,7 @@ class Api::V1::DeletionsControllerTest < ActionController::TestCase
       context "ON DELETE to create for an already deleted gem" do
         setup do
           Deletion.create!(user: @user, version: @v1)
-          delete :create, gem_name: @rubygem.to_param, version: @v1.number
+          delete :create, params: { gem_name: @rubygem.to_param, version: @v1.number }
         end
         should respond_with :unprocessable_entity
         should "not re-record the deletion" do
@@ -145,21 +145,21 @@ class Api::V1::DeletionsControllerTest < ActionController::TestCase
 
       context "ON PUT to destroy for version 0.1.0" do
         setup do
-          put :destroy, gem_name: @rubygem.to_param, version: @v1.number
+          put :destroy, params: { gem_name: @rubygem.to_param, version: @v1.number }
         end
         should respond_with :gone
       end
 
       context "ON PUT to destroy for version 0.1.2 and platform x86-darwin-10" do
         setup do
-          put :destroy, gem_name: @rubygem.to_param, version: @v3.number, platform: @v3.platform
+          put :destroy, params: { gem_name: @rubygem.to_param, version: @v3.number, platform: @v3.platform }
         end
         should respond_with :gone
       end
 
       context "ON PUT to destroy for version 0.1.1" do
         setup do
-          put :destroy, gem_name: @rubygem.to_param, version: @v2.number
+          put :destroy, params: { gem_name: @rubygem.to_param, version: @v2.number }
         end
         should respond_with :gone
       end
