@@ -31,7 +31,7 @@ class Api::V1::DownloadsControllerTest < ActionController::TestCase
   end
 
   def get_show(version, format = 'json')
-    get :show, id: version.full_name, format: format
+    get :show, params: { id: version.full_name }, format: format
   end
 
   def self.should_respond_to(format, to_meth = :to_s)
@@ -78,7 +78,7 @@ class Api::V1::DownloadsControllerTest < ActionController::TestCase
 
   context "on GET to show for an unknown gem" do
     setup do
-      get :show, id: "rials", format: 'json'
+      get :show, params: { id: "rials" }, format: 'json'
     end
 
     should "return a 404" do
