@@ -4,7 +4,7 @@ class ProfilesControllerTest < ActionController::TestCase
   context "for a user that doesn't exist" do
     should "throw a not found" do
       assert_raise ActiveRecord::RecordNotFound do
-        get :show, id: "unknown"
+        get :show, params: { id: "unknown" }
       end
     end
   end
@@ -24,33 +24,27 @@ class ProfilesControllerTest < ActionController::TestCase
           end
         end.reverse
 
-        get :show, id: @user.handle
+        get :show, params: { id: @user.handle }
       end
 
       should respond_with :success
-      should render_template :show
-      should "assign the last 10 most downloaded gems" do
-        assert_equal @rubygems[0..9], assigns[:rubygems]
-      end
-      should "assign the extra gems you own" do
-        assert_equal [@rubygems.last], assigns[:extra_rubygems]
+      should "display all gems of user" do
+        11.times { |i| assert page.has_content? @rubygems[i].name }
       end
     end
 
     context "on GET to show with handle" do
       setup do
-        get :show, id: @user.handle
+        get :show, params: { id: @user.handle }
       end
 
       should respond_with :success
-      should render_template :show
     end
 
     context "on GET to show with id" do
-      setup { get :show, id: @user.id }
+      setup { get :show, params: { id: @user.id } }
 
       should respond_with :success
-      should render_template :show
       should "render Email link" do
         assert page.has_content?("Email Me")
         assert page.has_selector?("a[href='mailto:#{@user.email}']")
@@ -60,11 +54,10 @@ class ProfilesControllerTest < ActionController::TestCase
     context "on GET to show when hide email" do
       setup do
         @user.update(hide_email: true)
-        get :show, id: @user.id
+        get :show, params: { id: @user.id }
       end
 
       should respond_with :success
-      should render_template :show
       should "not render Email link" do
         refute page.has_content?("Email Me")
         refute page.has_selector?("a[href='mailto:#{@user.email}']")
@@ -75,7 +68,6 @@ class ProfilesControllerTest < ActionController::TestCase
       setup { get :edit }
 
       should respond_with :success
-      should render_template :edit
     end
 
     context "on PUT to update" do
@@ -84,7 +76,7 @@ class ProfilesControllerTest < ActionController::TestCase
           @handle = "john_m_doe"
           @user = create(:user, handle: "johndoe")
           sign_in_as(@user)
-          put :update, user: { handle: @handle }
+          put :update, params: { user: { handle: @handle } }
         end
 
         should respond_with :redirect
@@ -102,7 +94,7 @@ class ProfilesControllerTest < ActionController::TestCase
           @hide_email = true
           @user = create(:user, handle: "johndoe")
           sign_in_as(@user)
-          put :update, user: { handle: @handle, hide_email: @hide_email }
+          put :update, params: { user: { handle: @handle, hide_email: @hide_email } }
         end
 
         should respond_with :redirect
