@@ -11,11 +11,10 @@ class RubygemsControllerTest < ActionController::TestCase
       setup do
         @owners = [@user, create(:user)]
         @rubygem = create(:rubygem, owners: @owners, number: "1.0.0")
-        get :show, id: @rubygem.to_param
+        get :show, params: { id: @rubygem.to_param }
       end
 
       should respond_with :success
-      should render_template :show
       should "renders owner gems overview links" do
         @owners.each do |owner|
           assert page.has_selector?("a[href='#{profile_path(owner.display_id)}']")
@@ -28,7 +27,7 @@ class RubygemsControllerTest < ActionController::TestCase
         @owners = [@user, create(:user)]
         @rubygem = create(:rubygem, owners: @owners, number: "1.0.0")
         @rubygem.linkset = nil
-        get :show, id: @rubygem.to_param
+        get :show, params: { id: @rubygem.to_param }
       end
 
       should respond_with :success
@@ -40,11 +39,10 @@ class RubygemsControllerTest < ActionController::TestCase
     context "On GET to show for another user's gem" do
       setup do
         @rubygem = create(:rubygem)
-        get :show, id: @rubygem.to_param
+        get :show, params: { id: @rubygem.to_param }
       end
 
       should respond_with :success
-      should render_template :show
       should "not render edit link" do
         refute page.has_selector?("a[href='#{edit_rubygem_path(@rubygem)}']")
       end
@@ -53,11 +51,10 @@ class RubygemsControllerTest < ActionController::TestCase
     context "On GET to show for this user's gem" do
       setup do
         @rubygem = create(:rubygem, owners: [@user], number: "1.0.0")
-        get :show, id: @rubygem.to_param
+        get :show, params: { id: @rubygem.to_param }
       end
 
       should respond_with :success
-      should render_template :show
       should "render edit link" do
         assert page.has_selector?("a[href='#{edit_rubygem_path(@rubygem)}']")
       end
@@ -68,7 +65,7 @@ class RubygemsControllerTest < ActionController::TestCase
         @rubygem = create(:rubygem)
         create(:version, rubygem: @rubygem)
         create(:subscription, rubygem: @rubygem, user: @user)
-        get :show, id: @rubygem.to_param
+        get :show, params: { id: @rubygem.to_param }
       end
 
       should respond_with :success
@@ -81,7 +78,7 @@ class RubygemsControllerTest < ActionController::TestCase
       setup do
         @rubygem = create(:rubygem)
         create(:version, rubygem: @rubygem)
-        get :show, id: @rubygem.to_param
+        get :show, params: { id: @rubygem.to_param }
       end
 
       should respond_with :success
@@ -93,11 +90,10 @@ class RubygemsControllerTest < ActionController::TestCase
     context "On GET to edit for this user's gem" do
       setup do
         @rubygem = create(:rubygem, owners: [@user], number: "1.0.0")
-        get :edit, id: @rubygem.to_param
+        get :edit, params: { id: @rubygem.to_param }
       end
 
       should respond_with :success
-      should render_template :edit
       should "render form" do
         assert page.has_selector?("form")
         assert page.has_selector?("input#linkset_code")
@@ -113,7 +109,7 @@ class RubygemsControllerTest < ActionController::TestCase
       setup do
         @other_user = create(:user)
         @rubygem = create(:rubygem, owners: [@other_user], number: "1.0.0")
-        get :edit, id: @rubygem.to_param
+        get :edit, params: { id: @rubygem.to_param }
       end
       should respond_with :redirect
       should redirect_to('the homepage') { root_url }
@@ -125,13 +121,15 @@ class RubygemsControllerTest < ActionController::TestCase
         @url = "https://github.com/qrush/gemcutter"
         @rubygem = create(:rubygem, owners: [@user], number: "1.0.0")
         put :update,
-          id: @rubygem.to_param,
-          linkset: {
-            code: @url,
-            docs: 'http://docs.com',
-            wiki: 'http://wiki.com',
-            mail: 'http://mail.com',
-            bugs: 'http://bugs.com'
+          params: {
+            id: @rubygem.to_param,
+            linkset: {
+              code: @url,
+              docs: 'http://docs.com',
+              wiki: 'http://wiki.com',
+              mail: 'http://mail.com',
+              bugs: 'http://bugs.com'
+            }
           }
       end
       should respond_with :redirect
@@ -158,10 +156,9 @@ class RubygemsControllerTest < ActionController::TestCase
       setup do
         @rubygem = create(:rubygem, owners: [@user], number: "1.0.0")
         @url = "totally not a url"
-        put :update, id: @rubygem.to_param, linkset: { code: @url }
+        put :update, params: { id: @rubygem.to_param, linkset: { code: @url } }
       end
       should respond_with :success
-      should render_template :edit
       should "not update linkset" do
         assert_not_equal @url, Rubygem.last.linkset.code
       end
@@ -183,7 +180,6 @@ class RubygemsControllerTest < ActionController::TestCase
     end
 
     should respond_with :success
-    should render_template :index
     should "render links" do
       @gems.each do |g|
         assert page.has_content?(g.name)
@@ -232,10 +228,9 @@ class RubygemsControllerTest < ActionController::TestCase
       @gems = (1..3).map { |n| create(:rubygem, name: "agem#{n}") }
       @zgem = create(:rubygem, name: "zeta")
       create(:version, rubygem: @zgem)
-      get :index, letter: "z"
+      get :index, params: { letter: "z" }
     end
     should respond_with :success
-    should render_template :index
     should "render links" do
       assert page.has_content?(@zgem.name)
       assert page.has_selector?("a[href='#{rubygem_path(@zgem)}']")
@@ -250,11 +245,10 @@ class RubygemsControllerTest < ActionController::TestCase
         gem
       end
       create(:rubygem, name: "zeta")
-      get :index, letter: "asdf"
+      get :index, params: { letter: "asdf" }
     end
 
     should respond_with :success
-    should render_template :index
     should "render links" do
       @gems.each do |g|
         assert page.has_content?(g.name)
@@ -267,11 +261,10 @@ class RubygemsControllerTest < ActionController::TestCase
     setup do
       @latest_version = create(:version, created_at: 1.minute.ago)
       @rubygem = @latest_version.rubygem
-      get :show, id: @rubygem.to_param
+      get :show, params: { id: @rubygem.to_param }
     end
 
     should respond_with :success
-    should render_template :show
     should "render info about the gem" do
       assert page.has_content?(@rubygem.name)
       assert page.has_content?(@latest_version.number)
@@ -288,17 +281,17 @@ class RubygemsControllerTest < ActionController::TestCase
     end
     should "render plural licenses header for other than one license" do
       @latest_version.update_attributes(licenses: nil)
-      get :show, id: @rubygem.to_param
+      get :show, params: { id: @rubygem.to_param }
       assert page.has_content?("Licenses")
 
       @latest_version.update_attributes(licenses: ["MIT", "GPL-2"])
-      get :show, id: @rubygem.to_param
+      get :show, params: { id: @rubygem.to_param }
       assert page.has_content?("Licenses")
     end
 
     should "render singular license header for one line license" do
       @latest_version.update_attributes(licenses: ["MIT"])
-      get :show, id: @rubygem.to_param
+      get :show, params: { id: @rubygem.to_param }
       assert page.has_content?("License")
       assert page.has_no_content?("Licenses")
     end
@@ -312,11 +305,10 @@ class RubygemsControllerTest < ActionController::TestCase
         create(:version, number: "1.9.9", rubygem: @rubygem, created_at: 1.minute.ago),
         create(:version, number: "1.9.9.rc4", rubygem: @rubygem, created_at: 2.days.ago)
       ]
-      get :show, id: @rubygem.to_param
+      get :show, params: { id: @rubygem.to_param }
     end
 
     should respond_with :success
-    should render_template :show
     should "render info about the gem" do
       assert page.has_content?(@rubygem.name)
       assert page.has_content?(@versions[0].number)
@@ -344,9 +336,8 @@ class RubygemsControllerTest < ActionController::TestCase
       @rubygem = version.rubygem
     end
     context 'when signed out' do
-      setup { get :show, id: @rubygem.to_param }
+      setup { get :show, params: { id: @rubygem.to_param } }
       should respond_with :success
-      should render_template :show
       should "render info about the gem" do
         assert page.has_content?("This gem is not currently hosted on RubyGems.org")
         assert page.has_no_content?('Versions')
@@ -357,7 +348,7 @@ class RubygemsControllerTest < ActionController::TestCase
         @user = create(:user)
         sign_in_as @user
         create(:subscription, user: @user, rubygem: @rubygem)
-        get :show, id: @rubygem.to_param
+        get :show, params: { id: @rubygem.to_param }
       end
       should "have a visible unsubscribe link" do
         assert page.has_selector?("a[style='display:inline-block']", text: 'Unsubscribe')
@@ -368,10 +359,9 @@ class RubygemsControllerTest < ActionController::TestCase
   context "On GET to show for a gem with no versions" do
     setup do
       @rubygem = create(:rubygem)
-      get :show, id: @rubygem.to_param
+      get :show, params: { id: @rubygem.to_param }
     end
     should respond_with :success
-    should render_template :show
     should "render info about the gem" do
       assert page.has_content?("This gem is not currently hosted on RubyGems.org.")
     end
@@ -384,11 +374,10 @@ class RubygemsControllerTest < ActionController::TestCase
       @development = create(:dependency, :development, version: @version)
       @runtime     = create(:dependency, :runtime,     version: @version)
 
-      get :show, id: @version.rubygem.to_param
+      get :show, params: { id: @version.rubygem.to_param }
     end
 
     should respond_with :success
-    should render_template :show
     should "show runtime dependencies and development dependencies" do
       assert page.has_content?(@runtime.rubygem.name)
       assert page.has_content?(@development.rubygem.name)
@@ -401,11 +390,10 @@ class RubygemsControllerTest < ActionController::TestCase
 
       @unresolved = create(:dependency, :unresolved, version: @version)
 
-      get :show, id: @version.rubygem.to_param
+      get :show, params: { id: @version.rubygem.to_param }
     end
 
     should respond_with :success
-    should render_template :show
     should "show unresolved dependencies" do
       assert page.has_content?(@unresolved.name)
     end
@@ -424,11 +412,10 @@ class RubygemsControllerTest < ActionController::TestCase
       @missing_dependency.rubygem.update_column(:name, 'missing')
       @missing_dependency.update_column(:rubygem_id, nil)
 
-      get :show, id: @version.rubygem.to_param
+      get :show, params: { id: @version.rubygem.to_param }
     end
 
     should respond_with :success
-    should render_template :show
     should "show only dependencies that have rubygem" do
       assert page.has_content?(@runtime.rubygem.name)
       assert page.has_no_content?('1.2.0')
@@ -440,11 +427,10 @@ class RubygemsControllerTest < ActionController::TestCase
       @version = create(:version)
       @runtime = create(:dependency, :runtime, version: @version)
       @runtime.rubygem.update_column(:name, 'foo>0.1.1')
-      get :show, id: @version.rubygem.to_param
+      get :show, params: { id: @version.rubygem.to_param }
     end
 
     should respond_with :success
-    should render_template :show
     should "show runtime dependencies and development dependencies" do
       assert page.has_content?(@runtime.rubygem.name)
     end
@@ -452,7 +438,7 @@ class RubygemsControllerTest < ActionController::TestCase
 
   context "On GET to show for nonexistent gem" do
     setup do
-      get :show, id: "blahblah"
+      get :show, params: { id: "blahblah" }
     end
 
     should respond_with :not_found
@@ -463,7 +449,7 @@ class RubygemsControllerTest < ActionController::TestCase
       setup do
         @rubygem = create(:rubygem)
         create(:version, rubygem: @rubygem)
-        get :show, id: @rubygem.to_param
+        get :show, params: { id: @rubygem.to_param }
       end
 
       should respond_with :success
@@ -478,7 +464,7 @@ class RubygemsControllerTest < ActionController::TestCase
     context "On GET to edit" do
       setup do
         @rubygem = create(:rubygem)
-        get :edit, id: @rubygem.to_param
+        get :edit, params: { id: @rubygem.to_param }
       end
       should respond_with :redirect
       should redirect_to('the homepage') { root_url }
@@ -487,7 +473,7 @@ class RubygemsControllerTest < ActionController::TestCase
     context "On PUT to update" do
       setup do
         @rubygem = create(:rubygem)
-        put :update, id: @rubygem.to_param, linkset: {}
+        put :update, params: { id: @rubygem.to_param, linkset: {} }
       end
       should respond_with :redirect
       should redirect_to('the homepage') { root_url }
