@@ -9,7 +9,6 @@ class SearchesControllerTest < ActionController::TestCase
     setup { get :show }
 
     should respond_with :success
-    should render_template :show
     should "see no results" do
       refute page.has_content?("Results")
     end
@@ -20,11 +19,10 @@ class SearchesControllerTest < ActionController::TestCase
       @sinatra = create(:rubygem, name: "sinatra")
       assert_nil @sinatra.versions.most_recent
       assert @sinatra.reload.versions.count.zero?
-      get :show, query: "sinatra"
+      get :show, params: { query: "sinatra" }
     end
 
     should respond_with :success
-    should render_template :show
   end
 
   context 'on GET to show with search parameters' do
@@ -35,11 +33,10 @@ class SearchesControllerTest < ActionController::TestCase
       create(:version, rubygem: @sinatra)
       create(:version, rubygem: @sinatra_redux)
       create(:version, rubygem: @brando)
-      get :show, query: "sinatra"
+      get :show, params: { query: "sinatra" }
     end
 
     should respond_with :success
-    should render_template :show
     should "see sinatra on the page in the results" do
       assert page.has_content?(@sinatra.name)
       assert page.has_selector?("a[href='#{rubygem_path(@sinatra)}']")
@@ -66,11 +63,10 @@ class SearchesControllerTest < ActionController::TestCase
       @brando.index_document
       Rubygem.__elasticsearch__.refresh_index!
       @request.cookies['new_search'] = 'true'
-      get :show, query: 'sinatra'
+      get :show, params: { query: 'sinatra' }
     end
 
     should respond_with :success
-    should render_template :show
     should "see sinatra on the page in the results" do
       page.assert_text(@sinatra.name)
       page.assert_selector("a[href='#{rubygem_path(@sinatra)}']")
@@ -92,7 +88,7 @@ class SearchesControllerTest < ActionController::TestCase
     setup do
       @sinatra = create(:rubygem, name: "sinatra")
       create(:version, rubygem: @sinatra)
-      get :show, query: "sinatra"
+      get :show, params: { query: "sinatra" }
     end
 
     should respond_with :redirect
@@ -101,11 +97,10 @@ class SearchesControllerTest < ActionController::TestCase
 
   context 'on GET to show with non string search parameter' do
     setup do
-      get :show, query: { foo: "bar" }
+      get :show, params: { query: { foo: "bar" } }
     end
 
     should respond_with :success
-    should render_template :show
   end
 
   context 'on GET to show with search parameters and no results' do
@@ -121,11 +116,10 @@ class SearchesControllerTest < ActionController::TestCase
       @brando.index_document
       Rubygem.__elasticsearch__.refresh_index!
       @request.cookies['new_search'] = 'true'
-      get :show, query: "sinatre"
+      get :show, params: { query: "sinatre" }
     end
 
     should respond_with :success
-    should render_template :show
     should "see sinatra on the page in the suggestions" do
       page.assert_text('Maybe you mean')
       assert page.find('.search__suggestions').has_content?(@sinatra.name)
@@ -151,7 +145,7 @@ class SearchesControllerTest < ActionController::TestCase
       requires_toxiproxy
       Toxiproxy[:elasticsearch].down do
         @request.cookies['new_search'] = 'true'
-        get :show, query: 'sinatra'
+        get :show, params: { query: 'sinatra' }
         assert_response :success
         assert page.has_content?('Advanced search is currently unavailable')
         assert page.has_content?('Displaying')
