@@ -12,7 +12,7 @@ class UsersControllerTest < ActionController::TestCase
   context "on POST to create" do
     context "when email and password are given" do
       should "create a user" do
-        post :create, user: { email: 'foo@bar.com', password: 'secret' }
+        post :create, params: { user: { email: 'foo@bar.com', password: 'secret' } }
         assert User.find_by(email: 'foo@bar.com')
       end
     end
@@ -27,12 +27,12 @@ class UsersControllerTest < ActionController::TestCase
 
     context "when extra parameters given" do
       should "create a user if parameters are ok" do
-        post :create, user: { email: 'foo@bar.com', password: 'secret', handle: 'foo' }
+        post :create, params: { user: { email: 'foo@bar.com', password: 'secret', handle: 'foo' } }
         assert_equal "foo", User.where(email: 'foo@bar.com').pluck(:handle).first
       end
 
       should "create a user but dont assign not valid parameters" do
-        post :create, user: { email: 'foo@bar.com', password: 'secret', api_key: 'nonono' }
+        post :create, params: { user: { email: 'foo@bar.com', password: 'secret', api_key: 'nonono' } }
         assert_not_equal "nonono", User.where(email: 'foo@bar.com').pluck(:api_key).first
       end
     end
