@@ -7,15 +7,14 @@ class GemsTest < ActionDispatch::IntegrationTest
   end
 
   test "gem page with a non valid HTTP_ACCEPT header" do
-    get rubygem_path(@rubygem), nil,
-      'HTTP_ACCEPT' => 'application/mercurial-0.1'
+    get rubygem_path(@rubygem), headers: { 'HTTP_ACCEPT' => 'application/mercurial-0.1' }
     assert page.has_content? "1.0.0"
   end
 
   test "gems page with atom format" do
     get rubygems_path(format: :atom)
     assert_response :success
-    assert_equal :atom, response.content_type.symbol
+    assert_equal 'application/atom+xml', response.content_type
     assert page.has_content? "sandworm"
   end
 
@@ -23,8 +22,8 @@ class GemsTest < ActionDispatch::IntegrationTest
     get rubygem_path(@rubygem, as: @user.id)
     assert page.has_css?('a#subscribe')
 
-    post rubygem_subscription_path(@rubygem, as: @user.id), nil,
-      'HTTP_ACCEPT' => 'application/javascript'
+    post rubygem_subscription_path(@rubygem, as: @user.id),
+      headers: { 'HTTP_ACCEPT' => 'application/javascript' }
 
     assert_match(/\("\.toggler"\)\.toggle\(\)/, @response.body)
     assert_equal @user.subscribed_gems.first, @rubygem
@@ -36,15 +35,15 @@ class GemsTest < ActionDispatch::IntegrationTest
     get rubygem_path(@rubygem, as: @user.id)
     assert page.has_css?('a#unsubscribe')
 
-    delete rubygem_subscription_path(@rubygem, as: @user.id), nil,
-      'HTTP_ACCEPT' => 'application/javascript'
+    delete rubygem_subscription_path(@rubygem, as: @user.id),
+      headers: { 'HTTP_ACCEPT' => 'application/javascript' }
     assert_match(/\("\.toggler"\)\.toggle\(\)/, @response.body)
   end
 
   test "versions with atom format" do
     create(:version, rubygem: @rubygem)
     get rubygem_versions_path(@rubygem, format: :atom)
-    assert_equal :atom, response.content_type.symbol
+    assert_equal 'application/atom+xml', response.content_type
     assert page.has_content? "sandworm"
   end
 
