@@ -11,8 +11,9 @@ class OwnerTest < ActionDispatch::IntegrationTest
   end
 
   test "adding an owner" do
-    post api_v1_rubygem_owners_path(@rubygem), { email: @other_user.email },
-      "HTTP_AUTHORIZATION" => @user.api_key
+    post api_v1_rubygem_owners_path(@rubygem),
+      params: { email: @other_user.email },
+      headers: { "HTTP_AUTHORIZATION" => @user.api_key }
     assert_response :success
 
     get rubygem_path(@rubygem)
@@ -22,8 +23,9 @@ class OwnerTest < ActionDispatch::IntegrationTest
 
   test "removing an owner" do
     create(:ownership, user: @other_user, rubygem: @rubygem)
-    delete api_v1_rubygem_owners_path(@rubygem), { email: @other_user.email },
-      "HTTP_AUTHORIZATION" => @user.api_key
+    delete api_v1_rubygem_owners_path(@rubygem),
+      params: { email: @other_user.email },
+      headers: { "HTTP_AUTHORIZATION" => @user.api_key }
 
     get rubygem_path(@rubygem)
     assert page.has_selector?("a[alt='#{@user.handle}']")
@@ -33,8 +35,9 @@ class OwnerTest < ActionDispatch::IntegrationTest
   test "transferring ownership" do
     create(:ownership, user: @other_user, rubygem: @rubygem)
 
-    delete api_v1_rubygem_owners_path(@rubygem), { email: @user.email },
-      "HTTP_AUTHORIZATION" => @user.api_key
+    delete api_v1_rubygem_owners_path(@rubygem),
+      params: { email: @user.email },
+      headers: { "HTTP_AUTHORIZATION" => @user.api_key }
 
     get rubygem_path(@rubygem)
     refute page.has_selector?("a[alt='#{@user.handle}']")
@@ -42,12 +45,14 @@ class OwnerTest < ActionDispatch::IntegrationTest
   end
 
   test "adding ownership without permission" do
-    post api_v1_rubygem_owners_path(@rubygem), { email: @other_user.email },
-      "HTTP_AUTHORIZATION" => @other_user.api_key
+    post api_v1_rubygem_owners_path(@rubygem),
+      params: { email: @other_user.email },
+      headers: { "HTTP_AUTHORIZATION" => @other_user.api_key }
     assert_response :unauthorized
 
-    delete api_v1_rubygem_owners_path(@rubygem), { email: @other_user.email },
-      "HTTP_AUTHORIZATION" => @other_user.api_key
+    delete api_v1_rubygem_owners_path(@rubygem),
+      params: { email: @other_user.email },
+      headers: { "HTTP_AUTHORIZATION" => @other_user.api_key }
     assert_response :unauthorized
   end
 end
